Bug 756: Added duplcation of nodes in TemplateTree

+++ branches/MAC-20100311-1_0-release-branch/SAS/OTB/OTB/src/nl/astron/lofar/sas/otb/panels/TemplateMaintenancePanel.java
@@ -227,11 +227,26 @@ public class TemplateMaintenancePanel extends javax.swing.JPanel
         logger.debug("Trigger: " + evt.getActionCommand());
         if (evt.getActionCommand().equals("Delete")) {
             //Check  if the selected node isn't a leaf
-            if (itsSelectedNode != null && !itsSelectedNode.leaf) {
+            if (itsSelectedNode != null && !itsSelectedNode.leaf && itsSelectedNode.instances <= 1 ) {
                 if (JOptionPane.showConfirmDialog(this, "Are you sure you want to delete this node ?", "Delete Tree", JOptionPane.YES_NO_OPTION, JOptionPane.WARNING_MESSAGE) == JOptionPane.YES_OPTION) {
                     try {
                         if (OtdbRmi.getRemoteMaintenance().deleteNode(itsSelectedNode)) {
                             logger.debug("Node + children deleted");
+
+//                            check diff between node duplication like beam en like duplicates
+                            itsMainFrame.ToDo();
+                            // We have to find the defaultNode for this deleted node and decrease the number of instances
+                            Vector<jOTDBnode> aList = OtdbRmi.getRemoteMaintenance().getItemList(itsTreeID, itsSelectedNode.name);
+                            Iterator<jOTDBnode> it = aList.iterator ();
+                            while (it.hasNext ()) {
+                                jOTDBnode aN = it.next ();
+                                if (aN.index == -1) {
+                                    // found defaultNode
+                                    Integer k = aN.instances-1;
+                                    aN.instances=k.shortValue();
+                                    OtdbRmi.getRemoteMaintenance().saveNode(aN);
+                                }
+                            }
                             setNewRootNode();
                         }
                     } catch (RemoteException ex) {
@@ -241,12 +256,8 @@ public class TemplateMaintenancePanel extends javax.swing.JPanel
             }
         } else if (evt.getActionCommand().equals("Duplicate")) {
 
-            //!!!!!  if duplicating a node the default node.instances need to be set to +1 instance!!!!!!
-            itsMainFrame.ToDo();
-
-
-            //Check  if the selected node isn't a leaf
-            if (itsSelectedNode != null && !itsSelectedNode.leaf) {
+            //Check  if the selected node isn't a leaf and it is a deafult node (index = -1)
+            if (itsSelectedNode != null && !itsSelectedNode.leaf && itsSelectedNode.index==-1) {
                 String answer = JOptionPane.showInputDialog(this, "What is the index for the new subtree?", "Enter indexNumber", JOptionPane.QUESTION_MESSAGE);
                 if (answer != null || !answer.equals("")) {
                     short idx = Integer.valueOf(answer).shortValue();
@@ -258,7 +269,13 @@ public class TemplateMaintenancePanel extends javax.swing.JPanel
                         int aN = OtdbRmi.getRemoteMaintenance().dupNode(itsTreeID, itsSelectedNode.nodeID(), idx);
                         if (aN > 0) {
                             logger.debug("Node duplicated");
-                            setNewRootNode();
+                            // defaultNode.instances needs 2b set to highest instance
+                           if (idx > itsSelectedNode.index) {
+                               Short newInst=idx;
+                               itsSelectedNode.instances=newInst;
+                               OtdbRmi.getRemoteMaintenance().saveNode(itsSelectedNode);
+                               setNewRootNode();
+                           }
                         } else {
                             logger.debug("Node duplication failed");
                         }
@@ -389,7 +406,7 @@ public class TemplateMaintenancePanel extends javax.swing.JPanel
         int savedSelection = jTabbedPane1.getSelectedIndex();
         logger.debug("ChangeSelection for node: " + aNode.name);
         itsSelectedNode = aNode;
-        jOTDBparam aParam = null;
+
 
         changeSelection(aNode);
 
@@ -399,12 +416,23 @@ public class TemplateMaintenancePanel extends javax.swing.JPanel
 
         } else {
             int[] selectedRows = treePanel.getSelectedRows();
-            if (selectedRows == null || selectedRows.length <= 0 || selectedRows[0] == 0 || aNode.leaf) {
+            if (selectedRows == null || selectedRows.length <= 0 || selectedRows[0] == 0 || aNode.leaf ) {
                 buttonPanel1.setButtonEnabled("Duplicate", false);
                 buttonPanel1.setButtonEnabled("Delete", false);
             } else {
-//                buttonPanel1.setButtonEnabled("Duplicate",true);
-                buttonPanel1.setButtonEnabled("Delete", true);
+                // only duplication possible on DefaultTrees (index = -1)
+                if (aNode.index == -1) {
+                    buttonPanel1.setButtonEnabled("Duplicate",true);
+                } else {
+                    buttonPanel1.setButtonEnabled("Duplicate",false);
+                }
+                // only deletion possible when no instances left make install
+
+//                if (aNode.index == -1 && aNode.instances == 1 ) {
+//                    buttonPanel1.setButtonEnabled("Delete", true);
+//                } else {
+//                    buttonPanel1.setButtonEnabled("Delete", false);
+//                }
             }
         }
         if (savedSelection > -1 && savedSelection < jTabbedPane1.getComponentCount()) {
@@ -424,9 +452,9 @@ public class TemplateMaintenancePanel extends javax.swing.JPanel
 
         } else {
             buttonPanel1.addButton("Delete");
+            buttonPanel1.setButtonEnabled("Delete", false);
             buttonPanel1.addButton("Duplicate");
             buttonPanel1.addButton("Exit");
-            buttonPanel1.setButtonEnabled("Duplicate", false);
         }
